feat: ProductDetailMain에서 타입 분리
