+++ src/widgets/product/lib/checkSameProductInSelected.ts
@@ -1,0 +1,18 @@
+import { SelectedProduct } from "../model";
+
+
+
+
+
+export const checkSameProductInSelected = (
+  products: SelectedProduct[],
+  colorId: number,
+  sizedId: number
+) => {
+  for (var product of products) {
+    if (colorId === product.colorId && sizedId === product.sizeId) {
+      return true;
+    }
+  }
+  return false;
+};
