+++ src/widgets/product/lib/index.ts
@@ -1,0 +1,2 @@
+export * from './checkSameProductInSelected';
+export * from './sumSelectedProducts';
