+++ src/widgets/product/model/types.ts
@@ -1,0 +1,14 @@
+
+
+
+export interface SelectedProduct {
+  key: number;
+  productId: number;
+  colorId: number;
+  colorName: string;
+  sizeId: number;
+  sizeName: string;
+  price: number;
+  stock: number;
+  quantity: number;
+}
